Improves robustness of bibfile cite key extraction.
Formatted bib entries with leading whitespace before the cite key
cause problems because the whitespace is extracted alongside the key.
This means output of tools such as bibtool is incorrectly processed.
For example:

@Article{	  foo,
  title		= {Bar},
  author	= {Doe, J.},
  journal	= {Nonexistent Journal},
  volume	= {1},
  number	= {1},
  pages		= {1--2},
  year		= {2012},
  publisher	= {Nonexistent publisher}
}

This would lead to the cite key being read as '	  foo'. This change
updates the extraction regex to not capture any optional leading
whitespace, resulting in just 'foo'.